Fix theta_H0_range usage by adding a key to dict
This seems to be necessary for 'theta_H0_range' to be implemented correctly. The H0 values need to be a list for the dict entry `more_config['cosmology_params']['theta_H0_range']`, not for `more_config['cosmology_params']['theta_H0_range']`, otherwise you end up passing a list to camb.set_cosmology instead of keyword arguments.