Added one point functionality to the two_point
This adds the functionality to the two_point.py which enables to include a one-point statistics to the cosmosis fits files and allows those to be included in the likelihood. It probably needs more testing and bug fixing, as I have probably not capture all edge cases when implementing  the additions. Due to the behaviour of one-point statistics, a lot was borrowed from tne n(z) functionality.